Replaced NSMapTable with CFMutableDictionaryRef for iOS compatibility
May break compatibility with other users of the SnoizeMIDI framework.
Haven't renamed variables from 'mapTable', now they're no longer technically map tables.

Also made some minor changes to types of format strings causing errors with 64-bit, and other issues with deprecated methods.

Haven’t yet created the iOS Static Library target.

+++ Frameworks/SnoizeMIDI/SMSysExSendRequest.m
@@ -86,7 +86,7 @@ NSString *SMSysExSendRequestFinishedNotification = @"SMSysExSendRequestFinishedN
     
     status = MIDISendSysex(&request);
     if (status) {
-        NSLog(@"MIDISendSysex() returned error: %ld", status);
+        NSLog(@"MIDISendSysex() returned error: %ld", (long)status);
         [self release];
     }
 }
